server/requirements.txt
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
feedparser==6.0.2
    # via -r requirements.in
flask-cors==3.0.10
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-cors
idna==2.10
    # via requests
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
python-dateutil==2.8.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
sgmllib3k==1.0.0
    # via feedparser
six==1.15.0
    # via
    #   flask-cors
    #   python-dateutil
urllib3==1.26.4
    # via requests
werkzeug==1.0.1
    # via flask
